pyproject.toml
[project]
name = "nova-mlops-lab"
version = "0.1.0"
description = "OpenStack-first MLOps job runner with a LightSpeed-style interface"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "rich>=13.7",
]

[project.optional-dependencies]
train = ["torch>=2.1", "torchvision>=0.16"]
dev = ["pytest>=8", "ruff>=0.5", "mypy>=1.10", "types-PyYAML"]
openstack = ["openstacksdk>=3.0.0"]


[project.scripts]
nova-mlops = "nova_mlops.cli:app"

[tool.ruff]
line-length = 100

dependencies = [
  "typer",
  "rich",
  "pyyaml",
  "openstacksdk",
]
